feat(profile): Add logout functionality and minor aside refactor
The logout used to be an anchor tag pointing to an invalid path, this code adds a way to make it now
a button that triggers a call to logout the user from our session service provider without leave the
app. Ultimately this code also introduce minor updates to type definition and methods naming.

re #3